Seed initialized

+++ db/migrate/20201208202112_create_specs.rb
@@ -3,7 +3,7 @@ class CreateSpecs < ActiveRecord::Migration[6.0]
     create_table :specs do |t|
       t.string :company_name
       t.string :model
-      t.integer :price
+      t.float :price
       t.string :status
       t.string :body_type
       t.string :fuel_type
@@ -12,7 +12,7 @@ class CreateSpecs < ActiveRecord::Migration[6.0]
       t.string :maximum_torque
       t.string :fuel_tank_capacity
       t.string :number_of_gears
-      t.integer :zero_to_100
+      t.float :zero_to_100
       t.timestamps
     end
   end
